Rename not() to no()
The name not() was influenced by the JavaScript operator ! and the function
not() in XSLT. However, using a similar but different name makes its use
clearer when reading code: !no(value) is now read as "not no value" instead
of "not (operator) not (function) value". Also, the not() function in XSLT
converts node sets to booleans, a kind of type transformation that the use
of no() intends to avoid in JavaScript.