Move assert.js to tests/ folder
The rationale is to differentiate easily functions used for unit tests,
test() and assert(), from functions under test or used as helpers for the
unit tests, located in the parent folder.

This move does not prevent from using both test() and assert() separately
in other contexts, but highlights their role in current project.